requirements.txt
streamlit
opencv-python-headless
numpy
tensorflow
pillow
sqlite3
pandas

